+++ src/test/java/io/fabric8/maven/docker/access/ContainerCreateConfigTest.java
@@ -3,6 +3,9 @@ package io.fabric8.maven.docker.access;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
+import io.fabric8.maven.docker.config.Arguments;
+import io.fabric8.maven.docker.config.HealthCheckConfiguration;
+import io.fabric8.maven.docker.config.HealthCheckMode;
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -11,13 +14,11 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import io.fabric8.maven.docker.util.JsonFactory;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /*
  *
@@ -51,12 +52,12 @@ class ContainerCreateConfigTest {
         Assertions.assertNotNull(env);
         Assertions.assertEquals(6, env.size());
         List<String> envAsString = convertToList(env);
-        Assertions.assertTrue(envAsString.contains("JAVA_OPTS=-Xmx512m"));
-        Assertions.assertTrue(envAsString.contains("TEST_SERVICE=SECURITY"));
-        Assertions.assertTrue(envAsString.contains("EXTERNAL_ENV=TRUE"));
-        Assertions.assertTrue(envAsString.contains("TEST_HTTP_ADDR=${docker.container.consul.ip}"));
-        Assertions.assertTrue(envAsString.contains("TEST_CONSUL_IP=+${docker.container.consul.ip}:8080"));
-        Assertions.assertTrue(envAsString.contains("TEST_CONSUL_IP_WITHOUT_DELIM=${docker.container.consul.ip}:8225"));
+        assertTrue(envAsString.contains("JAVA_OPTS=-Xmx512m"));
+        assertTrue(envAsString.contains("TEST_SERVICE=SECURITY"));
+        assertTrue(envAsString.contains("EXTERNAL_ENV=TRUE"));
+        assertTrue(envAsString.contains("TEST_HTTP_ADDR=${docker.container.consul.ip}"));
+        assertTrue(envAsString.contains("TEST_CONSUL_IP=+${docker.container.consul.ip}:8080"));
+        assertTrue(envAsString.contains("TEST_CONSUL_IP_WITHOUT_DELIM=${docker.container.consul.ip}:8225"));
     }
 
     @Test
@@ -83,7 +84,7 @@ class ContainerCreateConfigTest {
 
         JsonObject volumes = (JsonObject) JsonFactory.newJsonObject(cc.toJson()).get("Volumes");
         Assertions.assertEquals(1, volumes.size());
-        Assertions.assertTrue(volumes.has(expectedContainerPath));
+        assertTrue(volumes.has(expectedContainerPath));
     }
 
 
@@ -102,7 +103,7 @@ class ContainerCreateConfigTest {
         JsonArray env = getEnvArray(cc);
         Assertions.assertEquals(2, env.size());
         List<String> envAsString = convertToList(env);
-        Assertions.assertTrue(envAsString.contains("EXTERNAL_ENV=TRUE"));
+        assertTrue(envAsString.contains("EXTERNAL_ENV=TRUE"));
     }
 
     @Test
@@ -116,6 +117,70 @@ class ContainerCreateConfigTest {
     void platform() {
         ContainerCreateConfig cc= new ContainerCreateConfig("testImage", "linux/arm64");
         Assertions.assertEquals("linux/arm64",  cc.getPlatform());
+    }
+
+    @Test
+    void testHealthCheckIsFullyPresent() {
+        ContainerCreateConfig cc = new ContainerCreateConfig("testImage");
+        HealthCheckConfiguration healthCheckConfiguration = new HealthCheckConfiguration.Builder()
+                .cmd(new Arguments(Arrays.asList("CMD-SHELL", "some command 2>&1")))
+                .startPeriod("1s")
+                .interval("500ms")
+                .timeout("2s")
+                .retries(20)
+                .mode(HealthCheckMode.cmd)
+                .build();
+        cc.healthcheck(healthCheckConfiguration);
+        JsonObject config = JsonFactory.newJsonObject(cc.toJson());
+        JsonObject healthCheck = (JsonObject) config.get("Healthcheck");
+        Assertions.assertEquals(2, healthCheck.get("Test")
+                .getAsJsonArray().size());
+        Assertions.assertEquals("some command 2>&1", healthCheck.get("Test")
+                .getAsJsonArray().get(1).getAsString());
+        Assertions.assertEquals(20, healthCheck.get("Retries").getAsInt());
+        Assertions.assertEquals(500000000, healthCheck.get("Interval").getAsInt());
+        Assertions.assertEquals(1000000000, healthCheck.get("StartPeriod").getAsInt());
+        Assertions.assertEquals(2000000000, healthCheck.get("Timeout").getAsInt());
+    }
+
+    @Test
+    void testHealthCheckIsInherited() {
+        ContainerCreateConfig cc = new ContainerCreateConfig("testImage");
+        HealthCheckConfiguration healthCheckConfiguration = new HealthCheckConfiguration.Builder()
+                .cmd(new Arguments(Arrays.asList()))
+                .startPeriod("1s")
+                .mode(HealthCheckMode.cmd)
+                .build();
+        cc.healthcheck(healthCheckConfiguration);
+        JsonObject config = JsonFactory.newJsonObject(cc.toJson());
+        JsonObject healthCheck = (JsonObject) config.get("Healthcheck");
+        Assertions.assertEquals(0, healthCheck.get("Test")
+                .getAsJsonArray().size());
+        Assertions.assertNull(healthCheck.get("Retries"));
+        Assertions.assertNull(healthCheck.get("Interval"));
+        Assertions.assertEquals(1000000000, healthCheck.get("StartPeriod").getAsInt());
+        Assertions.assertNull(healthCheck.get("Timeout"));
+    }
+
+    @Test
+    void testHealthCheckIsDisabled() {
+        ContainerCreateConfig cc = new ContainerCreateConfig("testImage");
+        HealthCheckConfiguration healthCheckConfiguration = new HealthCheckConfiguration.Builder()
+                .cmd(new Arguments(Arrays.asList("NONE")))
+                .mode(HealthCheckMode.none)
+                .retries(2)
+                .build();
+        cc.healthcheck(healthCheckConfiguration);
+        JsonObject config = JsonFactory.newJsonObject(cc.toJson());
+        JsonObject healthCheck = (JsonObject) config.get("Healthcheck");
+        Assertions.assertEquals(1, healthCheck.get("Test")
+                .getAsJsonArray().size());
+        Assertions.assertEquals("NONE", healthCheck.get("Test")
+                .getAsJsonArray().get(0).getAsString());
+        Assertions.assertNull(healthCheck.get("Retries"));
+        Assertions.assertNull(healthCheck.get("Interval"));
+        Assertions.assertNull(healthCheck.get("StartPeriod"));
+        Assertions.assertNull(healthCheck.get("Timeout"));
     }
 
     private JsonArray getEnvArray(ContainerCreateConfig cc) {